android/foundation: Persist read permission for URI
On recent Android versions, attempting to read a "content://" resource
by its URI value would throw a SecurityException; as the necessary
permission is apparently a property of the Uri class. To overcome this
rather odd constraint, we ask ContentResolver to persist the grant so
that it would no longer be attached to the specific Uri instance but
rather the app itself as a whole.