android/journal3: Stop relying on global event hub for activity comms
We faced issues with our previous method of using the global event
source for cross-activity communication. It caused state restoration
problems when the starting activity got destroyed by the system. To fix
this, we now use the Activity Results API, treating the started activity
as an implementation detail, similar to media selection flow.