android/foundation: Suppress false warnings from Lint
Lint raised a false-positive warning about the usage of forEach() within
the function. This is a recurring bug from Lint wherein it fails to read
the metadata of newer Kotlin binaries.

It is possible that this has been addressed in newer versions of Lint,
but since it means we'd have to upgrade the entire Android Gradle Plugin
to get it, we'll just suppress the warning for now.