kotlin/geography: Respect the cache while finding by ID
Workaround the limitation of doing direct lookup to HERE's services.
Start caching past search results and regard the data whenever we do
an ID-based find. This is done to resolve a crash that would happen
when the user selects a place that was not a part of the original
iterator (e.g., search results).